Note about missing config/releases.exs
Came across an issue that caused the firmware to build but not boot while upgrading to Elixir 1.9+ releases. I was working through the docs and `mix` gave an error about a missing `config/releases.exs` file so I created an empty file, problem solved. 

Except it causes an empty release configuration to be created which prevent BEAM from starting. It gives an error like: 

> 'Elixir.ArgumentError',message=><<"could not fetch environment variable \"RELEASE SYS_CONFIG\" because it is not set" ...

Not sure if it's just me but figured it'd be helpful to add a note in the docs.